FractalDimensions.jl

FractalDimensions — Module

FractalDimensions.jl

A Julia package that estimates various definitions of fractal dimension from data. It can be used as a standalone package, or as part of DynamicalSystems.jl.

To install it, run import Pkg; Pkg.add("FractalDimensions").

All further information is provided in the documentation, which you can either find online or build locally by running the docs/make.jl file.

Previously, this package was part of ChaosTools.jl.

Citation

If you use this package in a publication, please cite the paper below:

@ARTICLE{FractalDimensions.jl,
  title     = "Estimating the fractal dimension: a comparative review and open
               source implementations",
  author    = "Datseris, George and Kottlarz, Inga and Braun, Anton P and
               Parlitz, Ulrich",
  publisher = "arXiv",
  year      =  2021,
  doi = {10.48550/ARXIV.2109.05937},
  url = {https://arxiv.org/abs/2109.05937},
}

Introduction

Note

This package is accompanying a review paper on the fractal dimension: https://arxiv.org/abs/2109.05937. The paper is continuing the discussion of chapter 5 of Nonlinear Dynamics, Datseris & Parlitz, Springer 2022.

There are numerous methods that one can use to calculate a so-called "dimension" of a dataset which in the context of dynamical systems is called the Fractal dimension. In the majority of cases, computing a fractal dimension means estimating the scaling behaviour of some quantity as a size/scale increases. In the Fractal dimension example below, one finds the scaling of the entropy of the histogram of some data, versus the width of the bins of the histogram. In this case, it approximately holds $ H \approx -\Delta\log(\varepsilon) $ for bin width $\varepsilon$. The scaling of many other quantities can be estimated as well, such as the correlation sum, the Higuchi length, or others provided here.

To actually find $\Delta$, one needs to find a linearly scaling region in the graph $H$ vs. $\log(\varepsilon)$ and estimate its slope. Hence, identifying a linear region is central to estimating a fractal dimension. That is why, the section Linear scaling regions is of central importance for this documentation.

Fractal dimension example

In this simplest example we will calculate the fractal dimension of the chaotic attractor of the Hénon map (for default parameters).

using DynamicalSystemsBase: DeterministicIteratedMap, trajectory
using CairoMakie

henon_rule(x, p, n) = SVector(1.0 - p[1]*x[1]^2 + x[2], p[2]*x[1])
u0 = zeros(2)
p0 = [1.4, 0.3]
henon = DeterministicIteratedMap(henon_rule, u0, p0)

X, t = trajectory(henon, 100_000; Ttr = 100)
scatter(X[:, 1], X[:, 2]; color = ("black", 0.01), markersize = 4)

Our goal is to compute entropies of the histogram of the above plot for many different partition sizes (bin widths) ε. Computing entropies is the job of ComplexityMeasures.jl, but the two relevant names (entropy, ValueHistogram) are re-exported by FractalDimensions.jl.

using FractalDimensions
ες = 2 .^ (-15:0.5:5) # semi-random guess
Hs = [entropy(ValueHistogram(ε), X) for ε in ες]

xs = @. -log2(ες) # must use same base as `entropy`!!!
scatterlines(xs, Hs; axis = (ylabel = L"H_1", xlabel = L"-\log (\epsilon)"))

The slope of the linear scaling region of the above plot is the generalized dimension (of order q = 1) for the attractor of the Hénon map.

Given that we see the plot, we can estimate where the linear scaling region starts and ends. However, we can use the function linear_region to get an estimate of the result as well. First let's visualize what it does, as it uses linear_regions.

lrs, slopes = linear_regions(xs, Hs, tol = 0.25)
fig = Figure()
ax = Axis(fig[1,1]; ylabel = L"H_1", xlabel = L"-\log (\epsilon)")
for r in lrs
    scatterlines!(ax, xs[r], Hs[r])
end
fig

The linear_region function finds, and computes the slope of, the largest region:

Δ = linear_region(xs, Hs)[2]

This result is an approximation of the information dimension (because we used q = 1) of the attractor.

The whole above pipeline we went through is bundled in generalized_dim. Similar pipeline is done by grassberger_proccacia_dim and many other functions.

Be wary when using `xxxxx_dim`

As stated clearly by the documentation strings, all pre-made dimension estimating functions (ending in _dim) perform a lot of automated steps, each having its own heuristic choices for function default values. They are more like convenient bundles with on-average good defaults, rather than precise functions. You should be careful when considering the validity of the returned number!

Linear scaling regions

And other utilities, especially linreg, used in both [generalized_dim] and grassberger_dim.

FractalDimensions.linear_regions — Function

linear_regions(x, y; dxi::Int = 1, tol = 0.25) -> (lrs, tangents)

Identify regions where the curve y(x) is linear, by scanning the x-axis every dxi indices sequentially (e.g. at x[1] to x[5], x[5] to x[10], x[10] to x[15] and so on if dxi=5).

If the slope (calculated via linear regression) of a region of width dxi is approximatelly equal to that of the previous region, within tolerance tol, then these two regions belong to the same linear region.

Return the indices of x that correspond to the linear regions, lrs, and the correct tangents at each region (obtained via a second linear regression at each accumulated region). lrs is hence a vector of UnitRanges.

FractalDimensions.linear_region — Function

linear_region(x, y; kwargs...) -> (region, slope)

Call linear_regions and identify and return the largest linear region (a UnitRange of the indices of x) and its corresponding slope.

The keywords dxi, tol are propagated as-is to linear_regions. The keyword ignore_saturation = true ignores saturation that (sometimes) happens at the start and end of the curve y(x), where the curve flattens. The keyword sat = 0.01 decides what saturation is (while abs(y[i]-y[i+1])<sat we are in a saturation regime).

The keyword warning = true prints a warning if the linear region is less than 1/3 of the available x-axis.

FractalDimensions.linreg — Function

linreg(x, y) -> a, b

Perform a linear regression to find the best coefficients so that the curve: z = a + b*x has the least squared error with y.

FractalDimensions.estimate_boxsizes — Function

estimate_boxsizes(X::AbstractStateSpaceSet; kwargs...) → εs

Return k exponentially spaced values: εs = base .^ range(lower + w, upper + z; length = k), that are a good estimate for sizes ε that are used in calculating a Fractal Dimension. It is strongly recommended to standardize input dataset before using this function.

Let d₋ be the minimum pair-wise distance in X, d₋ = dminimum_pairwise_distance(X). Let d₊ be the average total length of X, d₊ = mean(ma - mi) with mi, ma = minmaxima(X). Then lower = log(base, d₋) and upper = log(base, d₊). Because by default w=1, z=-1, the returned sizes are an order of mangitude larger than the minimum distance, and an order of magnitude smaller than the maximum distance.

Keywords

w = 1, z = -1, k = 16 : as explained above.

base = MathConstants.e : the base used in the log function.

warning = true: Print some warnings for bad estimates.

autoexpand = true: If the final estimated range does not cover at least 2 orders of magnitude, it is automatically expanded by setting w -= we and z -= ze. You can set different default values to the keywords we = w, ze = z.

FractalDimensions.minimum_pairwise_distance — Function

minimum_pairwise_distance(X::StateSpaceSet, kdtree = dimension(X) < 10, metric = Euclidean())

Return min_d, min_pair: the minimum pairwise distance of all points in the dataset, and the corresponding point pair. The third argument is a switch of whether to use KDTrees or a brute force search.

Generalized (entropy) dimension

Based on the definition of the Generalized entropy (genentropy), one can calculate an appropriate dimension, called generalized dimension:

FractalDimensions.generalized_dim — Function

generalized_dim(X::StateSpaceSet [, sizes]; q = 1, base = 2) -> Δ_q

Return the q order generalized dimension of X, by calculating its histogram-based Rényi entropy for each ε ∈ sizes.

The case of q = 0 is often called "capacity" or "box-counting" dimension, while q = 1 is the "information" dimension.

Description

The returned dimension is approximated by the (inverse) power law exponent of the scaling of the Renyi entropy $H_q$, versus the box size ε, where ε ∈ sizes:

\[H_q \approx -\Delta_q\log_{b}(\varepsilon)\]

$H_q$ is calculated using ComplexityMeasures: Renyi, ValueHistogram, entropy, i.e., by doing a histogram of the data with a given box size.

Calling this function performs a lot of automated steps:

A vector of box sizes is decided by calling sizes = estimate_boxsizes(dataset), if sizes is not given.

For each element of sizes the appropriate entropy is calculated as
H = [entropy(Renyi(; q, base), ValueHistogram(ε), data) for ε ∈ sizes]
Let x = -log.(sizes).

The curve H(x) is decomposed into linear regions, using linear_regions(x, h).

The biggest linear region is chosen, and a fit for the slope of that region is performed using the function linear_region, which does a simple linear regression fit using linreg. This slope is the return value of generalized_dim.

By doing these steps one by one yourself, you can adjust the keyword arguments given to each of these function calls, refining the accuracy of the result. The source code of this function is only 3 lines of code.

FractalDimensions.molteno_dim — Function

molteno_dim(X::AbstractStateSpaceSet; k0::Int = 10, q = 1.0, base = 2)

Return an estimate of the generalized_dim of X using the algorithm by [Molteno1993]. This function is a simple utilization of the probabilities estimated by molteno_boxing so see that function for more details. Here the entropy of the probabilities is computed at each size, and a line is fitted in the entropy vs log(size) graph, just like in generalized_dim.

FractalDimensions.molteno_boxing — Function

molteno_boxing(X::AbstractStateSpaceSet; k0::Int = 10) → (probs, εs)

Distribute X into boxes whose size is halved in each step, according to the algorithm by [Molteno1993]. Division stops if the average number of points per filled box falls below the threshold k0.

Return probs, a vector of Probabilities of finding points in boxes for different box sizes, and the corresponding box sizes εs. These outputs are used in molteno_dim.

Description

Project the data onto the whole interval of numbers that is covered by UInt64. The projected data is distributed into boxes whose size decreases by factor 2 in each step. For each box that contains more than one point 2^D new boxes are created where D is the dimension of the data.

The process of dividing the data into new boxes stops when the number of points over the number of filled boxes falls below k0. The box sizes εs are calculated and returned together with the probs.

This algorithm is faster than the traditional approach of using ValueHistogram(ε::Real), but it is only suited for low dimensional data since it divides each box into 2^D new boxes if D is the dimension. For large D this leads to low numbers of box divisions before the threshold is passed and the divison stops. This results to a low number of data points to fit the dimension to and thereby a poor estimate.

Correlation sum based dimension

FractalDimensions.grassberger_proccacia_dim — Function

grassberger_proccacia_dim(X::AbstractStateSpaceSet, εs = estimate_boxsizes(data); kwargs...)

Use the method of Grassberger and Proccacia[Grassberger1983], and the correction by Theiler[Theiler1986], to estimate the correlation dimension Δ_C of X.

This function does something extremely simple:

cm = correlationsum(data, εs; kwargs...)
Δ_C = linear_region(log2.(sizes), log2.(cm))[2]

i.e. it calculates correlationsum for various radii and then tries to find a linear region in the plot of the log of the correlation sum versus log(ε).

See correlationsum for the available keywords. See also takens_best_estimate, boxassisted_correlation_dim.

FractalDimensions.correlationsum — Function

correlationsum(X, ε::Real; w = 0, norm = Euclidean(), q = 2) → C_q(ε)

Calculate the q-order correlation sum of X (StateSpaceSet or timeseries) for a given radius ε and norm. They keyword show_progress = true can be used to display a progress bar for large X.

correlationsum(X, εs::AbstractVector; w, norm, q) → C_q(ε)

If εs is a vector, C_q is calculated for each ε ∈ εs more efficiently. Multithreading is also enabled over the available threads (Threads.nthreads()). The function boxed_correlationsum is typically faster if the dimension of X is small and if maximum(εs) is smaller than the size of X.

Keyword arguments

q = 2: order of the correlation sum

norm = Euclidean(): distance norm

w = 0: Theiler window

show_progress = true: display a progress bar

Description

The correlation sum is defined as follows for q=2:

\[C_2(\epsilon) = \frac{2}{(N-w)(N-w-1)}\sum_{i=1}^{N}\sum_{j=1+w+i}^{N} B(||X_i - X_j|| < \epsilon)\]

for as follows for q≠2

\[C_q(\epsilon) = \left[\frac{1}{\alpha} \sum_{i=w+1}^{N-w} \left[\sum_{j:|i-j| > w} B(||X_i - X_j|| < \epsilon)\right]^{q-1}\right]^{1/(q-1)}\]

where

\[\alpha = (N-2w)(N-2w-1)^{(q-1)}\]

with $N$ the length of X and $B$ gives 1 if its argument is true. w is the Theiler window.

See the article of Grassberger for the general definition [Grassberger2007] and the book "Nonlinear Time Series Analysis" [Kantz2003], Ch. 6, for a discussion around choosing best values for w, and Ch. 11.3 for the explicit definition of the q-order correlationsum. Note that the formula in 11.3 is incorrect, but corrected here, and also note that we immediatelly exponentiate $C_q$ to $1/(q-1)$, so that it scales exponentially as $C_q \propto \varepsilon ^\Delta_q$ versus the size $\varepsilon$.

Box-assisted version

FractalDimensions.boxassisted_correlation_dim — Function

boxassisted_correlation_dim(X::AbstractStateSpaceSet; kwargs...)

Use the box-assisted optimizations of [Bueno2007] to estimate the correlation dimension Δ_C of X.

This function does something extremely simple:

εs, Cs = boxed_correlationsum(X; kwargs...)
return linear_region(log2.(Cs), log2.(εs))[2]

and hence see boxed_correlationsum for more information and available keywords.

FractalDimensions.boxed_correlationsum — Function

boxed_correlationsum(X::AbstractStateSpaceSet, εs, r0 = maximum(εs); kwargs...) → Cs

Estimate the correlationsum for each size ε ∈ εs using an optimized algorithm that first distributes data into boxes of size r0, and then computes the correlation sum for each box and each neighboring box of each box. This method is much faster than correlationsum, provided that the box size r0 is significantly smaller than the attractor length. Good choices for r0 are estimate_r0_buenoorovio or estimate_r0_theiler.

boxed_correlationsum(X::AbstractStateSpaceSet; kwargs...) → εs, Cs

In this method the minimum inter-point distance and estimate_r0_buenoorovio of X are used to estimate suitable εs for the calculation, which are also returned.

Keyword arguments

q = 2 : The order of the correlation sum.

P = 2 : The prism dimension.

w = 0 : The Theiler window.

show_progress = false : Whether to display a progress bar for the calculation.

norm = Euclidean() : Distance norm.

Description

C_q(ε) is calculated for every ε ∈ εs and each of the boxes to then be summed up afterwards. The method of splitting the data into boxes was implemented according to Theiler[Theiler1987]. w is the Theiler window. P is the prism dimension. If P is unequal to the dimension of the data, only the first P dimensions are considered for the box distribution (this is called the prism-assisted version). By default P is 2, which is the version suggested by [Bueno2007]. Alternative for P is the prismdim_theiler. Note that only when P = dimension(X) the boxed version is guaranteed to be exact to the original correlationsum. For any other P, some point pairs that should have been included may be skipped due to having smaller distance in the remaining dimensions, but larger distance in the first P dimensions.

FractalDimensions.prismdim_theiler — Function

prismdim_theiler(X)

An algorithm to find the ideal choice of a prism dimension for boxed_correlationsum using Theiler's original suggestion.

FractalDimensions.estimate_r0_buenoorovio — Function

estimate_r0_buenoorovio(X::AbstractStateSpaceSet, P = 2) → r0, ε0

Estimate a reasonable size for boxing X, proposed by Bueno-Orovio and Pérez-García[Bueno2007], before calculating the correlation dimension as presented by Theiler[Theiler1983]. Return the size r0 and the minimum interpoint distance ε0 in the data.

If instead of boxes, prisms are chosen everything stays the same but P is the dimension of the prism. To do so the dimension ν is estimated by running the algorithm by Grassberger and Procaccia[Grassberger1983] with √N points where N is the number of total data points. An effective size ℓ of the attractor is calculated by boxing a small subset of size N/10 into boxes of sidelength r_ℓ and counting the number of filled boxes η_ℓ.

\[\ell = r_\ell \eta_\ell ^{1/\nu}\]

The optimal number of filled boxes η_opt is calculated by minimising the number of calculations.

\[\eta_\textrm{opt} = N^{2/3}\cdot \frac{3^\nu - 1}{3^P - 1}^{1/2}.\]

P is the dimension of the data or the number of edges on the prism that don't span the whole dataset.

Then the optimal boxsize $r_0$ computes as

\[r_0 = \ell / \eta_\textrm{opt}^{1/\nu}.\]

FractalDimensions.estimate_r0_theiler — Function

estimate_r0_theiler(X::AbstractStateSpaceSet) → r0, ε0

Estimate a reasonable size for boxing the data X before calculating the boxed_correlationsum proposed by Theiler[Theiler1987]. Return the boxing size r0 and minimum inter-point distance in X, ε0.

To do so the dimension is estimated by running the algorithm by Grassberger and Procaccia[Grassberger1983] with √N points where N is the number of total data points. Then the optimal boxsize $r_0$ computes as

\[r_0 = R (2/N)^{1/\nu}\]

where $R$ is the size of the chaotic attractor and $\nu$ is the estimated dimension.

Fixed mass correlation sum

FractalDimensions.fixedmass_correlation_dim — Function

fixedmass_correlation_dim(X [, max_j]; kwargs...)

Use the fixed mass algorithm for computing the correlation sum, and use the result to compute the correlation dimension Δ_M of X.

This function does something extremely simple:

rs, ys = fixedmass_correlationsum(X, args...; kwargs...)
linear_region(rs, ys)[2]

FractalDimensions.fixedmass_correlationsum — Function

fixedmass_correlationsum(X [, max_j]; metric = Euclidean(), M = length(X)) → rs, ys

A fixed mass algorithm for the calculation of the correlationsum, and subsequently a fractal dimension $\Delta$, with max_j the maximum number of neighbours that should be considered for the calculation.

By default max_j = clamp(N*(N-1)/2, 5, 32) with N the data length.

Keyword arguments

M defines the number of points considered for the averaging of distances, randomly subsampling them from X.

metric = Euclidean() is the distance metric.

start_j = 4 computes the equation below starting from j = start_j. Typically the first j values have not converged to the correct scaling of the fractal dimension.

Description

"Fixed mass" algorithms mean that instead of trying to find all neighboring points within a radius, one instead tries to find the max radius containing j points. A correlation sum is obtained with this constrain, and equivalently the mean radius containing k points. Based on this, one can calculate $\Delta$ approximating the information dimension. The implementation here is due to to [Grassberger1988], which defines

\[Ψ(j) - \log N \sim \Delta \times \overline{\log \left( r_{(j)}\right)}\]

where $\Psi(j) = \frac{\text{d} \log Γ(j)}{\text{d} j}$ is the digamma function, rs = $\overline{\log \left( r_{(j)}\right)}$ is the mean logarithm of a radius containing j neighboring points, and ys = $\Psi(j) - \log N$ ($N$ is the length of the data). The amount of neighbors found $j$ range from 2 to max_j. The numbers are also converted to base $2$ from base $e$.

$\Delta$ can be computed by using linear_region(rs, ys).

Takens best estimate

FractalDimensions.takens_best_estimate_dim — Function

takens_best_estimate_dim(X, εmax, metric = Chebyshev(), εmin = 0)

Use the "Takens' best estimate" [Takens1985][Theiler1988] method for estimating the correlation dimension.

The original formula is

\[\Delta_C \approx \frac{C(\epsilon_\text{max})}{\int_0^{\epsilon_\text{max}}(C(\epsilon) / \epsilon) \, d\epsilon}\]

where $C$ is the correlationsum and $\epsilon_\text{max}$ is an upper cutoff. Here we use the later expression

\[\Delta_C \approx - \frac{1}{\eta},\quad \eta = \frac{1}{(N-1)^*}\sum_{[i, j]^*}\log(||X_i - X_j|| / \epsilon_\text{max})\]

where the sum happens for all $i, j$ so that $i < j$ and $||X_i - X_j|| < \epsilon_\text{max}$. In the above expression, the bias in the original paper has already been corrected, as suggested in [Borovkova1999].

According to [Borovkova1999], introducing a lower cutoff εmin can make the algorithm more stable (no divergence), this option is given but defaults to zero.

If X comes from a delay coordinates embedding of a timseries x, a recommended value for $\epsilon_\text{max}$ is std(x)/4.

You may also use

Δ_C, Δu_C, Δl_C = FractalDimensions.takens_best_estimate(args...)

to obtain the upper and lower 95% confidence intervals. The intervals are estimated from the log-likelihood function by finding the values of Δ_C where the function has fallen by 2 from its maximum, see e.g. [Barlow] chapter 5.3.

Kaplan-Yorke dimension

FractalDimensions.kaplanyorke_dim — Function

kaplanyorke_dim(λs::AbstractVector)

Calculate the Kaplan-Yorke dimension, a.k.a. Lyapunov dimension[Kaplan1970] from the given Lyapunov exponents λs.

Description

The Kaplan-Yorke dimension is simply the point where cumsum(λs) becomes zero (interpolated):

\[ D_{KY} = k + \frac{\sum_{i=1}^k \lambda_i}{|\lambda_{k+1}|},\quad k = \max_j \left[ \sum_{i=1}^j \lambda_i > 0 \right].\]

If the sum of the exponents never becomes negative the function will return the length of the input vector.

Useful in combination with lyapunovspectrum from ChaosTools.jl.

Higuchi dimension

FractalDimensions.higuchi_dim — Function

higuchi_dim(x::AbstractVector [, ks])

Estimate the Higuchi dimension[Higuchi1988] of the graph of x.

Description

The Higuchi dimension is a number Δ ∈ [1, 2] that quantifies the roughness of the graph of the function x(t), assuming here that x is equi-sampled, like in the original paper.

The method estimates how the length of the graph increases as a function of the indices difference (which, in this context, is equivalent with differences in t). Specifically, we calculate the average length versus k as

\[L_m(k) = \frac{N-1}{\lfloor \frac{N-m}{k} floor k^2} \sum_{i=1}^{\lfloor \frac{N-m}{k} \rfloor} |X_N(m+ik)-X_N(m+(i-1)k)| \\ L(k) = \frac{1}{k} \sum_{m=1}^k L_m(k)\]

and then use linear_region in -log2.(k) vs log2.(L) as per usual when computing a fractal dimension.

The algorithm chooses default ks to be exponentially spaced in base-2, up to at most 2^8. A user can provide their own ks as a second argument otherwise.

Use FractalDimensions.higuchi_length(x, ks) to obtain $L(k)$ directly.

Extreme value value theory dimension

FractalDimensions.extremevaltheory_dim — Function

extremevaltheory_dim(X::StateSpaceSet, p::Real; kwargs...) → Δ

Convenience syntax that returns the mean of the local dimensions of extremevaltheory_dims_persistences, which approximates a fractal dimension of X using extreme value theory and quantile probability p.

FractalDimensions.extremevaltheory_dims_persistences — Function

extremevaltheory_dims_persistences(x::AbstractStateSpaceSet, p::Real; kwargs)

Return the local dimensions Δloc and the persistences θloc for each point in the given set for quantile probability p, according to the estimation done via extreme value theory [Lucarini2016] [Caby2018]. The computation is parallelized to available threads (Threads.nthreads()).

Keyword arguments

show_progress = true: displays a progress bar.

estimator = :mm: how to estimate the σ parameter of the Generalized Pareto Distribution. The local fractal dimension is 1/σ. The possible values are: :exp, :mm, as in estimate_gpd_parameters.

compute_persistence = true: whether to aso compute local persistences θloc (also called extremal index). If false, θloc are NaNs.

allocate_matrix = false: If true, the code calls a method that attempts to allocate an N×N matrix (N = length(X)) that stores the pairwise Euclidean distances. This method is faster due to optimizations of Distances.pairwise but will error if the computer does not have enough available memory for the matrix allocation.

Description

For each state space point $\mathbf{x}_i$ in X we compute $g_j = -\log(||\mathbf{x}_i - \mathbf{x}_j|| ) \; \forall j = 1, \ldots, N$ with $||\cdot||$ the Euclidean distance. Next, we choose an extreme quantile probability $p$ (e.g., 0.99) for the distribution of $g_j$. We compute $g_p$ as the $p$-th quantile of $g_j$. Then, we collect the exceedances of $g_j$, defined as $E = \{ g_j - g_p: g_j \ge g_p \}$, i.e., all values of $g_j$ larger or equal to $g_p$, also shifted by $g_p$. There are in total $n = N(1-q)$ values in $E$. According to extreme value theory, in the limit $N \to \infty$ the values $E$ follow a two-parameter Generalized Pareto Distribution (GPD) with parameters $\sigma,\xi$ (the third parameter $\mu$ of the GPD is zero due to the positive-definite construction of $E$). Within this extreme value theory approach, the local dimension $\Delta^{(E)}_i$ assigned to state space point $\textbf{x}_i$ is given by the inverse of the $\sigma$ parameter of the GPD fit to the data[Faranda2011], $\Delta^{(E)}_i = 1/\sigma$. $\sigma$ is estimated according to the estimator keyword.

FractalDimensions.extremevaltheory_local_dim_persistence — Function

extremevaltheory_local_dim_persistence(X::StateSpaceSet, ζ, p::Real; kw...)

Return the local values Δ, θ of the fractal dimension and persistence of X around a state space point ζ. p and kw are as in extremevaltheory_dims_persistences.

FractalDimensions.extremal_index_sueveges — Function

extremal_index_sueveges(y::AbstractVector, p)

Compute the extremal index θ of y through the Süveges formula for quantile probability p.

FractalDimensions.estimate_gpd_parameters — Function

estimate_gpd_parameters(X::AbstractVector{<:Real}, estimator::Symbol = :mm)

Estimate and return the parameters σ, ξ of a Generalized Pareto Distribution fit to X, assuming that minimum(X) == 0 and hence the parameter μ is 0 (if not, simply shift X by its minimum), according to the methods provided in [Flavio2023].

Optionally choose the estimator, which can be:

:exp: Assume the distribution is exponential instead of GP and get σ from mean of X and set ξ = 0.

mm: Standing for "method of moments", estimants are given by
\[\xi = (\bar{x}^2/s^2 - 1)/2, \quad \sigma = \bar{x}(\bar{x}^2/s^2 + 1)/2\]
with $\bar{x}$ the sample mean and $s^2$ the sample variance. This estimator only exists if the true distribution ξ value is < 0.5.

Theiler window

The Theiler window is a concept that is useful when finding neighbors in a dataset that is coming from the sampling of a continuous dynamical system. Itt tries to eliminate spurious "correlations" (wrongly counted neighbors) due to a potentially dense sampling of the trajectory. Typically a good choice for w coincides with the choice an optimal delay time, see DelayEmbeddings.estimate_delay, for any of the timeseries of the dataset.

For more details, see Chapter 5 of Nonlinear Dynamics, Datseris & Parlitz, Springer 2022.

StateSpaceSet reference

StateSpaceSets.StateSpaceSet — Type

StateSpaceSet{D, T} <: AbstractStateSpaceSet{D,T}

A dedicated interface for sets in a state space. It is an ordered container of equally-sized points of length D. Each point is represented by SVector{D, T}. The data are a standard Julia Vector{SVector}, and can be obtained with vec(ssset::StateSpaceSet). Typically the order of points in the set is the time direction, but it doesn't have to be.

When indexed with 1 index, StateSpaceSet is like a vector of points. When indexed with 2 indices it behaves like a matrix that has each of the columns be the timeseries of each of the variables. When iterated over, it iterates over its contained points. See description of indexing below for more.

StateSpaceSet also supports almost all sensible vector operations like append!, push!, hcat, eachrow, among others.

Description of indexing

In the following let i, j be integers, typeof(X) <: AbstractStateSpaceSet and v1, v2 be <: AbstractVector{Int} (v1, v2 could also be ranges, and for performance benefits make v2 an SVector{Int}).

X[i] == X[i, :] gives the ith point (returns an SVector)

X[v1] == X[v1, :], returns a StateSpaceSet with the points in those indices.

X[:, j] gives the jth variable timeseries (or collection), as Vector

X[v1, v2], X[:, v2] returns a StateSpaceSet with the appropriate entries (first indices being "time"/point index, while second being variables)

X[i, j] value of the jth variable, at the ith timepoint

Use Matrix(ssset) or StateSpaceSet(matrix) to convert. It is assumed that each column of the matrix is one variable. If you have various timeseries vectors x, y, z, ... pass them like StateSpaceSet(x, y, z, ...). You can use columns(dataset) to obtain the reverse, i.e. all columns of the dataset in a tuple.

StateSpaceSets.standardize — Function

standardize(d::StateSpaceSet) → r

Create a standardized version of the input set where each column is transformed to have mean 0 and standard deviation 1.

standardize(x::AbstractVector{<:Real}) = (x - mean(x))/std(x)

Molteno1993Molteno, T. C. A., Fast O(N) box-counting algorithm for estimating dimensions. Phys. Rev. E 48, R3263(R) (1993)

Grassberger1983Grassberger and Proccacia, Characterization of strange attractors, PRL 50 (1983)

Theiler1986Theiler, Spurious dimension from correlation algorithms applied to limited time-series data. Physical Review A, 34

Grassberger2007Peter Grassberger (2007) Grassberger-Procaccia algorithm. Scholarpedia, 2(5):3043.

Kantz2003Kantz, H., & Schreiber, T. (2003). Nonlinear Time Series Analysis, Cambridge University Press.

Bueno2007Bueno-Orovio and Pérez-García, Enhanced box and prism assisted algorithms for computing the correlation dimension. Chaos Solitons & Fractrals, 34(5)

Theiler1987Theiler, Efficient algorithm for estimating the correlation dimension from a set of discrete points. Physical Review A, 36

Bueno2007Bueno-Orovio and Pérez-García, Enhanced box and prism assisted algorithms for computing the correlation dimension. Chaos Solitons & Fractrals, 34(5)

Theiler1987Theiler, Efficient algorithm for estimating the correlation dimension from a set of discrete points. Physical Review A, 36

Grassberger1983Grassberger and Proccacia, Characterization of strange attractors, PRL 50 (1983)

Theiler1987Theiler, Efficient algorithm for estimating the correlation dimension from a set of discrete points. Physical Review A, 36

Grassberger1983Grassberger and Proccacia, Characterization of strange attractors, PRL 50 (1983)

Grassberger1988Peter Grassberger (1988) Finite sample Corrections to Entropy and Dimension Estimates, Physics Letters A 128(6-7)

Takens1985Takens, On the numerical determination of the dimension of an attractor, in: B.H.W. Braaksma, B.L.J.F. Takens (Eds.), Dynamical Systems and Bifurcations, in: Lecture Notes in Mathematics, Springer, Berlin, 1985, pp. 99–106.

Theiler1988Theiler, Lacunarity in a best estimator of fractal dimension. Physics Letters A, 133(4–5)

Borovkova1999Borovkova et al., Consistency of the Takens estimator for the correlation dimension. The Annals of Applied Probability, 9, 05 1999.

BarlowBarlow, R., Statistics - A Guide to the Use of Statistical Methods in the Physical Sciences. Vol 29. John Wiley & Sons, 1993

Kaplan1970J. Kaplan & J. Yorke, Chaotic behavior of multidimensional difference equations, Lecture Notes in Mathematics vol. 730, Springer (1979)

Higuchi1988Higuchi, Approach to an irregular time series on the basis of the fractal theory, Physica D: Nonlinear Phenomena (1988)

Lucarini2016Lucarini et al., Extremes and Recurrence in Dynamical Systems

Caby2018Caby et al., Physica D 400 132143

Süveges2007Süveges. 2007. Likelihood estimation of the extremal index. Extremes, 10.1-2, 41-55, doi: 10.1007/s10687-007-0034-2

Flavio2023Flavio et al., Stability of attractor local dimension estimates in non-Axiom A dynamical systems, preprint